Began migrating MAVlink wrapper functions to various modules throughout the directory
