+++ util/commands/mav_modes.py
@@ -1,0 +1,4 @@
+"""
+Functions for changing the operating mode of the AUV
+"""
+
